refactor(test): split test_sqlite_backend.py by backend concern (#494)
Split the 1222-line test_sqlite_backend.py into 6 focused files:

- test_sqlite_crud.py: init, create/read/update/delete, all/count/clear, exists/ping (27 tests)
- test_sqlite_inbox_outbox.py: inbox/outbox queues, enqueue_callback (17 tests)
- test_sqlite_rehydration.py: nested-object dehydration, rehydration, hydrate() (13 tests)
- test_sqlite_coercion.py: TestRehydrateFields, TestCoerceToSemanticClass (7 tests)
- test_sqlite_list.py: TestListMethod / list_objects() (5 tests)
- test_sqlite_find_case.py: find_case_by_short_id, find_case_by_report_id (9 tests)

Shared fixtures (dl, tmp_db_url, file_dl, scoped_dl, record_factory,
created_record) consolidated in test/adapters/driven/conftest.py. The dl
fixture is now in-memory (sqlite:///:memory:) matching the original local
override, and faster for unit tests.

All 3449 unit tests pass. All 6 new files are ≤300 lines.
