--- test/adapters/driven/test_sqlite_backend.py
@@ -1,1222 +1,0 @@
-#  Copyright (c) 2026 Carnegie Mellon University and Contributors.
-#  - see Contributors.md for a full list of Contributors
-#  - see ContributionInstructions.md for information on how you can Contribute to this project
-#  Vultron Multiparty Coordinated Vulnerability Disclosure Protocol Prototype is
-#  licensed under a MIT (SEI)-style license, please see LICENSE.md distributed
-#  with this Software or contact permission@sei.cmu.edu for full terms.
-#  Created, in part, with funding and support from the United States Government
-#  (see Acknowledgments file). This program may include and/or can make use of
-#  certain third party source code, object code, documentation and other files
-#  ("Third Party Software"). See LICENSE.md for more details.
-#  Carnegie Mellon®, CERT® and CERT Coordination Center® are registered in the
-#  U.S. Patent and Trademark Office by Carnegie Mellon University
-
-"""Tests for the SQLite/SQLModel-backed DataLayer adapter.
-
-Unit tests use an in-memory SQLite database (no file I/O required).
-Integration tests (marked ``@pytest.mark.integration``) use a real
-file-backed database via ``tmp_path``.
-"""
-
-import os
-
-import pytest
-
-from vultron.adapters.driven.db_record import Record
-from vultron.adapters.driven.datalayer_sqlite import SqliteDataLayer
-from vultron.wire.as2.factories import (
-    announce_log_entry_activity,
-    em_propose_embargo_activity,
-    rm_invite_to_case_activity,
-    rm_submit_report_activity,
-)
-from vultron.wire.as2.vocab.base.objects.activities.transitive import (
-    as_Accept,
-    as_Announce,
-    as_Invite,
-    as_Offer,
-)
-
-_ZERO_HASH: str = "0" * 64  # arbitrary hash for test chains
-
-# ---------------------------------------------------------------------------
-# Fixtures
-# ---------------------------------------------------------------------------
-
-
-@pytest.fixture
-def dl():
-    """In-memory SqliteDataLayer for unit tests."""
-    instance = SqliteDataLayer("sqlite:///:memory:")
-    yield instance
-    instance.clear_all()
-    instance.close()
-
-
-@pytest.fixture
-def tmp_db_url(tmp_path):
-    """SQLite URL pointing to a temporary file (for integration tests)."""
-    db_path = tmp_path / "test_sqlite.db"
-    return f"sqlite:///{db_path}"
-
-
-@pytest.fixture
-def file_dl(tmp_db_url):
-    """File-backed SqliteDataLayer for integration tests."""
-    instance = SqliteDataLayer(tmp_db_url)
-    yield instance
-    instance.clear_all()
-    instance.close()
-
-
-@pytest.fixture
-def record_factory():
-    def _make(id_="12345", type_="test_table", data_=None):
-        data = {"field": "value"} if data_ is None else data_
-        return Record(id_=id_, type_=type_, data_=data)
-
-    return _make
-
-
-@pytest.fixture
-def created_record(dl, record_factory):
-    rec = record_factory()
-    dl.create(rec)
-    return rec
-
-
-# ---------------------------------------------------------------------------
-# Initialisation
-# ---------------------------------------------------------------------------
-
-
-@pytest.mark.integration
-def test_database_initialization_creates_db_file(tmp_db_url):
-    """Creating a file-backed SqliteDataLayer creates the SQLite file."""
-    instance = SqliteDataLayer(tmp_db_url)
-    # Extract path from "sqlite:///path/to/file.db"
-    db_path = tmp_db_url.replace("sqlite:///", "")
-    assert os.path.exists(db_path)
-    instance.close()
-
-
-def test_database_initialization_in_memory():
-    """In-memory DataLayer can be created and is operational."""
-    instance = SqliteDataLayer("sqlite:///:memory:")
-    assert instance.ping()
-    instance.close()
-
-
-# ---------------------------------------------------------------------------
-# create / read / get
-# ---------------------------------------------------------------------------
-
-
-def test_create_inserts_record(dl, record_factory):
-    record = record_factory()
-    dl.create(record)
-    got = dl.get(record.type_, record.id_)
-    assert got is not None
-    assert got["id_"] == record.id_
-    assert got["type_"] == record.type_
-    assert got["data_"] == record.data_
-
-
-def test_create_raises_on_duplicate_id(dl, record_factory):
-    record = record_factory()
-    dl.create(record)
-    with pytest.raises(ValueError):
-        dl.create(record)
-
-
-def test_get_returns_none_for_nonexistent_type(dl):
-    assert dl.get("nonexistent_table", "no_id") is None
-
-
-def test_get_returns_none_for_missing_id(dl, record_factory):
-    record = record_factory()
-    dl.create(record)
-    assert dl.get(record.type_, "no_such_id") is None
-
-
-def test_read_returns_object(dl, record_factory):
-    from vultron.wire.as2.vocab.base.objects.object_types import as_Note
-
-    note = as_Note(content="Test Content")
-    dl.save(note)
-    result = dl.read(note.id_)
-    assert result is not None
-    assert result.id_ == note.id_  # type: ignore[union-attr]
-
-
-def test_read_returns_none_for_missing_id(dl):
-    assert dl.read("urn:uuid:does-not-exist") is None
-
-
-def test_read_supports_bare_uuid(dl):
-    from vultron.wire.as2.vocab.base.objects.object_types import as_Note
-
-    note = as_Note(content="Bare UUID test")
-    dl.save(note)
-    # Strip the urn:uuid: prefix
-    bare = note.id_.replace("urn:uuid:", "")
-    result = dl.read(bare)
-    assert result is not None
-    assert result.id_ == note.id_  # type: ignore[union-attr]
-
-
-# ---------------------------------------------------------------------------
-# update / save / delete
-# ---------------------------------------------------------------------------
-
-
-def test_update_updates_existing_record_data(dl, created_record):
-    rec = created_record
-    new_data = rec.data_.copy()
-    new_data["field"] = "new_value"
-    updated_record = Record(id_=rec.id_, type_=rec.type_, data_=new_data)
-
-    updated = dl.update(id_=updated_record.id_, record=updated_record)
-    assert updated
-    got = dl.get(updated_record.type_, updated_record.id_)
-    assert got is not None
-    assert got["data_"]["field"] == "new_value"
-
-
-def test_update_returns_false_for_non_existing_id(dl, record_factory):
-    non_existing = record_factory(id_="no_such_id")
-    updated = dl.update(id_=non_existing.id_, record=non_existing)
-    assert not updated
-
-
-def test_save_inserts_new_object(dl):
-    from vultron.wire.as2.vocab.base.objects.object_types import as_Note
-
-    note = as_Note(content="Save test")
-    dl.save(note)
-    result = dl.read(note.id_)
-    assert result is not None
-
-
-def test_save_overwrites_existing_object(dl):
-    from vultron.wire.as2.vocab.base.objects.object_types import as_Note
-
-    note = as_Note(content="Original content")
-    dl.save(note)
-    # Change the content and save again
-    note2 = as_Note(id_=note.id_, content="Updated content")
-    dl.save(note2)
-    result = dl.read(note.id_)
-    assert result is not None
-    assert result.content == "Updated content"  # type: ignore[union-attr]
-
-
-def test_delete_removes_record_and_returns_true(dl, record_factory):
-    record = record_factory()
-    dl.create(record)
-    deleted = dl.delete(record.type_, record.id_)
-    assert deleted
-    assert dl.get(record.type_, record.id_) is None
-
-
-def test_delete_returns_false_for_nonexistent_record(dl, record_factory):
-    record = record_factory()
-    assert not dl.delete(record.type_, record.id_)
-
-
-# ---------------------------------------------------------------------------
-# all / get_all / by_type / count_all
-# ---------------------------------------------------------------------------
-
-
-def test_all_returns_all_records_for_table(dl, record_factory):
-    record1 = record_factory(id_="id1", data_={"field": "value1"})
-    record2 = record_factory(id_="id2", data_={"field": "value2"})
-    dl.create(record1)
-    dl.create(record2)
-
-    all_records = dl.all("test_table")
-    assert len(all_records) == 2
-    ids = {rec.id_ for rec in all_records}  # type: ignore[union-attr]
-    assert "id1" in ids
-    assert "id2" in ids
-    for rec in all_records:
-        assert isinstance(rec, Record)
-
-
-def test_all_without_table_returns_dict_of_objects(dl):
-    from vultron.wire.as2.vocab.base.objects.object_types import as_Note
-
-    note = as_Note(content="All test")
-    dl.save(note)
-    result = dl.all()
-    assert isinstance(result, dict)
-    assert note.id_ in result
-
-
-def test_get_all_returns_list_of_dicts(dl, record_factory):
-    rec = record_factory()
-    dl.create(rec)
-    result = dl.get_all("test_table")
-    assert isinstance(result, list)
-    assert len(result) == 1
-    assert result[0]["id_"] == rec.id_
-
-
-def test_by_type_returns_dict(dl, record_factory):
-    rec = record_factory()
-    dl.create(rec)
-    result = dl.by_type("test_table")
-    assert isinstance(result, dict)
-    assert rec.id_ in result
-
-
-def test_count_all_returns_dict_with_counts(dl, record_factory):
-    rec = record_factory()
-    dl.create(rec)
-    counts = dl.count_all()
-    assert isinstance(counts, dict)
-    assert "_default" in counts
-    assert counts["test_table"] == 1
-
-
-# ---------------------------------------------------------------------------
-# clear_table / clear_all
-# ---------------------------------------------------------------------------
-
-
-def test_clear_table_removes_all_records(dl, record_factory):
-    record = record_factory()
-    dl.create(record)
-    dl.clear_table(record.type_)
-    assert dl.get(record.type_, record.id_) is None
-    assert len(dl.all(record.type_)) == 0
-
-
-def test_clear_all_removes_all_records(dl, record_factory):
-    record1 = record_factory(id_="id1", type_="table1")
-    record2 = record_factory(id_="id2", type_="table2")
-    dl.create(record1)
-    dl.create(record2)
-    dl.clear_all()
-    assert dl.get(record1.type_, record1.id_) is None
-    assert dl.get(record2.type_, record2.id_) is None
-
-
-# ---------------------------------------------------------------------------
-# exists / ping
-# ---------------------------------------------------------------------------
-
-
-def test_exists_returns_false_for_nonexistent_type(dl):
-    assert not dl.exists("nonexistent_table", "no_id")
-
-
-def test_exists_returns_true_after_create(dl, record_factory):
-    record = record_factory()
-    dl.create(record)
-    assert dl.exists(record.type_, record.id_)
-
-
-def test_exists_returns_false_for_missing_id(dl, record_factory):
-    record = record_factory()
-    dl.create(record)
-    assert not dl.exists(record.type_, "no_such_id")
-
-
-def test_exists_returns_false_after_delete(dl, record_factory):
-    record = record_factory()
-    dl.create(record)
-    dl.delete(record.type_, record.id_)
-    assert not dl.exists(record.type_, record.id_)
-
-
-def test_ping_returns_true(dl):
-    assert dl.ping() is True
-
-
-# ---------------------------------------------------------------------------
-# Inbox / outbox queue methods
-# ---------------------------------------------------------------------------
-
-
-def test_inbox_starts_empty(dl):
-    assert dl.inbox_list() == []
-
-
-def test_inbox_append_and_list(dl):
-    dl.inbox_append("https://example.org/activities/001")
-    assert "https://example.org/activities/001" in dl.inbox_list()
-
-
-def test_inbox_pop_fifo_order(dl):
-    dl.inbox_append("https://example.org/activities/001")
-    dl.inbox_append("https://example.org/activities/002")
-    assert dl.inbox_pop() == "https://example.org/activities/001"
-    assert len(dl.inbox_list()) == 1
-
-
-def test_inbox_pop_empty_returns_none(dl):
-    assert dl.inbox_pop() is None
-
-
-def test_outbox_starts_empty(dl):
-    assert dl.outbox_list() == []
-
-
-def test_outbox_append_and_list(dl):
-    dl.outbox_append("https://example.org/activities/sent-001")
-    assert "https://example.org/activities/sent-001" in dl.outbox_list()
-
-
-def test_outbox_pop_fifo_order(dl):
-    dl.outbox_append("https://example.org/activities/sent-001")
-    dl.outbox_append("https://example.org/activities/sent-002")
-    assert dl.outbox_pop() == "https://example.org/activities/sent-001"
-    assert len(dl.outbox_list()) == 1
-
-
-def test_outbox_pop_empty_returns_none(dl):
-    assert dl.outbox_pop() is None
-
-
-# ---------------------------------------------------------------------------
-# Nested-object dehydration integration tests
-# ---------------------------------------------------------------------------
-
-
-def test_activity_with_nested_object_stores_id_reference_not_inline_copy(dl):
-    """Storing an activity with a nested object writes only the nested ID."""
-    from vultron.wire.as2.vocab.objects.vulnerability_report import (
-        VulnerabilityReport,
-    )
-
-    report = VulnerabilityReport(
-        name="Test CVE",
-        content="A critical vulnerability",
-        attributed_to="https://example.org/finder",
-    )
-    offer = rm_submit_report_activity(
-        report,
-        "https://example.org/finder",
-        actor="https://example.org/finder",
-    )
-
-    dl.create(report)
-    dl.create(offer)
-
-    raw = dl.get(offer.type_, offer.id_)
-    assert raw is not None
-    stored_object_field = raw["data_"]["object_"]
-    assert isinstance(
-        stored_object_field, str
-    ), f"Expected object_ to be stored as ID string, got {type(stored_object_field)}"
-    assert stored_object_field == report.id_
-
-
-def test_activity_nested_object_retrievable_separately_after_dehydration(dl):
-    from vultron.wire.as2.vocab.objects.vulnerability_report import (
-        VulnerabilityReport,
-    )
-
-    report = VulnerabilityReport(
-        name="Test CVE 2",
-        content="Another vulnerability",
-        attributed_to="https://example.org/finder",
-    )
-    offer = rm_submit_report_activity(
-        report,
-        "https://example.org/finder",
-        actor="https://example.org/finder",
-    )
-
-    dl.create(report)
-    dl.create(offer)
-
-    retrieved_report = dl.read(report.id_)
-    assert retrieved_report is not None
-    assert retrieved_report.id_ == report.id_  # type: ignore[union-attr]
-
-
-def test_reading_activity_back_yields_expanded_nested_object(dl):
-    """After DL-REHYDRATE, dl.read() expands dehydrated object_ fields.
-
-    Previously ``dl.read()`` returned a string ID for the nested ``object_``
-    field; after the rehydration pipeline the full typed object is returned.
-    """
-    from vultron.wire.as2.vocab.objects.vulnerability_report import (
-        VulnerabilityReport,
-    )
-
-    report = VulnerabilityReport(
-        name="Test CVE 3",
-        content="Yet another vulnerability",
-        attributed_to="https://example.org/finder",
-    )
-    offer = rm_submit_report_activity(
-        report,
-        "https://example.org/finder",
-        actor="https://example.org/finder",
-    )
-
-    dl.create(report)
-    dl.create(offer)
-
-    retrieved_offer = dl.read(offer.id_)
-    assert retrieved_offer is not None
-    # object_ is now a fully-expanded VulnerabilityReport, not a string ID
-    assert isinstance(retrieved_offer.object_, VulnerabilityReport)  # type: ignore[union-attr]
-    assert retrieved_offer.object_.id_ == report.id_  # type: ignore[union-attr]
-
-
-def test_rehydration_restores_nested_object_from_datalayer(dl):
-    from vultron.wire.as2.rehydration import rehydrate
-    from vultron.wire.as2.vocab.objects.vulnerability_report import (
-        VulnerabilityReport,
-    )
-
-    report = VulnerabilityReport(
-        name="Test CVE 4",
-        content="One more vulnerability",
-        attributed_to="https://example.org/finder",
-    )
-    offer = rm_submit_report_activity(
-        report,
-        "https://example.org/finder",
-        actor="https://example.org/finder",
-    )
-
-    dl.create(report)
-    dl.create(offer)
-
-    stored_offer = dl.read(offer.id_)
-    assert stored_offer is not None
-    full_offer = rehydrate(stored_offer, dl=dl)
-    assert full_offer.object_ is not None  # type: ignore[union-attr]
-    assert full_offer.object_.id_ == report.id_  # type: ignore[union-attr]
-    assert full_offer.object_.type_ == "VulnerabilityReport"  # type: ignore[union-attr]
-
-
-def test_rehydration_does_not_mutate_stored_record(dl):
-    from vultron.wire.as2.rehydration import rehydrate
-    from vultron.wire.as2.vocab.objects.vulnerability_report import (
-        VulnerabilityReport,
-    )
-
-    report = VulnerabilityReport(
-        name="Test CVE 5",
-        content="Final vulnerability",
-        attributed_to="https://example.org/finder",
-    )
-    offer = rm_submit_report_activity(
-        report,
-        "https://example.org/finder",
-        actor="https://example.org/finder",
-    )
-
-    dl.create(report)
-    dl.create(offer)
-
-    stored_offer = dl.read(offer.id_)
-    assert stored_offer is not None
-    rehydrate(stored_offer, dl=dl)
-
-    raw = dl.get(offer.type_, offer.id_)
-    assert raw is not None
-    stored_object_field = raw["data_"]["object_"]
-    assert isinstance(stored_object_field, str)
-    assert stored_object_field == report.id_
-
-
-# ---------------------------------------------------------------------------
-# find_case_by_short_id tests
-# ---------------------------------------------------------------------------
-
-
-def test_find_case_by_short_id_with_http_url_case_id(dl):
-    from vultron.wire.as2.vocab.objects.vulnerability_case import (
-        VulnerabilityCase,
-    )
-
-    case = VulnerabilityCase(id_="https://example.org/api/v2/cases/demo-123")
-    dl.save(case)
-
-    result = dl.find_case_by_short_id("demo-123")
-    assert result is not None
-    assert isinstance(result, VulnerabilityCase)
-    assert result.id_ == case.id_
-
-
-def test_find_case_by_short_id_with_urn_case_id(dl):
-    from vultron.wire.as2.vocab.objects.vulnerability_case import (
-        VulnerabilityCase,
-    )
-
-    case = VulnerabilityCase(
-        id_="urn:uuid:11111111-2222-3333-4444-555555555555"
-    )
-    dl.save(case)
-
-    result = dl.find_case_by_short_id("11111111-2222-3333-4444-555555555555")
-    assert result is not None
-    assert isinstance(result, VulnerabilityCase)
-    assert result.id_ == case.id_
-
-
-def test_find_case_by_short_id_returns_none_when_ambiguous(dl):
-    from vultron.wire.as2.vocab.objects.vulnerability_case import (
-        VulnerabilityCase,
-    )
-
-    case1 = VulnerabilityCase(id_="https://org1.example/api/v2/cases/shared")
-    case2 = VulnerabilityCase(id_="https://org2.example/api/v2/cases/shared")
-    dl.save(case1)
-    dl.save(case2)
-
-    assert dl.find_case_by_short_id("shared") is None
-
-
-# ---------------------------------------------------------------------------
-# find_case_by_report_id tests
-# ---------------------------------------------------------------------------
-
-
-def test_find_case_by_report_id_returns_case_when_report_stored_as_string(dl):
-    from vultron.wire.as2.vocab.objects.vulnerability_case import (
-        VulnerabilityCase,
-    )
-    from vultron.wire.as2.vocab.objects.vulnerability_report import (
-        VulnerabilityReport,
-    )
-
-    report = VulnerabilityReport(
-        name="CVE-2025-001",
-        content="Test vulnerability",
-        attributed_to="https://example.org/finder",
-    )
-    case = VulnerabilityCase()
-    case.vulnerability_reports.append(report.id_)
-
-    dl.create(report)
-    dl.save(case)
-
-    result = dl.find_case_by_report_id(report.id_)
-    assert result is not None
-    assert isinstance(result, VulnerabilityCase)
-    assert result.id_ == case.id_
-
-
-def test_find_case_by_report_id_returns_case_when_report_stored_as_object(dl):
-    from vultron.wire.as2.vocab.objects.vulnerability_case import (
-        VulnerabilityCase,
-    )
-    from vultron.wire.as2.vocab.objects.vulnerability_report import (
-        VulnerabilityReport,
-    )
-
-    report = VulnerabilityReport(
-        name="CVE-2025-002",
-        content="Another vulnerability",
-        attributed_to="https://example.org/finder",
-    )
-    case = VulnerabilityCase()
-    case.vulnerability_reports.append(report)
-
-    dl.create(report)
-    dl.save(case)
-
-    result = dl.find_case_by_report_id(report.id_)
-    assert result is not None
-    assert isinstance(result, VulnerabilityCase)
-    assert result.id_ == case.id_
-
-
-def test_find_case_by_report_id_returns_none_when_not_found(dl):
-    from vultron.wire.as2.vocab.objects.vulnerability_case import (
-        VulnerabilityCase,
-    )
-    from vultron.wire.as2.vocab.objects.vulnerability_report import (
-        VulnerabilityReport,
-    )
-
-    report = VulnerabilityReport(
-        name="CVE-2025-003",
-        content="Unlinked vulnerability",
-        attributed_to="https://example.org/finder",
-    )
-    case = VulnerabilityCase()
-
-    dl.create(report)
-    dl.save(case)
-
-    result = dl.find_case_by_report_id(report.id_)
-    assert result is None
-
-
-def test_find_case_by_report_id_returns_none_when_no_cases(dl):
-    result = dl.find_case_by_report_id("urn:uuid:nonexistent-report")
-    assert result is None
-
-
-def test_find_case_by_report_id_returns_none_for_unknown_id(dl):
-    from vultron.wire.as2.vocab.objects.vulnerability_case import (
-        VulnerabilityCase,
-    )
-    from vultron.wire.as2.vocab.objects.vulnerability_report import (
-        VulnerabilityReport,
-    )
-
-    report = VulnerabilityReport(
-        name="CVE-2025-004",
-        content="Linked vulnerability",
-        attributed_to="https://example.org/finder",
-    )
-    other_report = VulnerabilityReport(
-        name="CVE-2025-005",
-        content="Unlinked vulnerability",
-        attributed_to="https://example.org/finder",
-    )
-    case = VulnerabilityCase()
-    case.vulnerability_reports.append(report.id_)
-
-    dl.create(report)
-    dl.create(other_report)
-    dl.save(case)
-
-    result = dl.find_case_by_report_id(other_report.id_)
-    assert result is None
-
-
-def test_find_case_by_report_id_returns_case_via_report_case_link(dl):
-    from vultron.core.models.report_case_link import VultronReportCaseLink
-    from vultron.wire.as2.vocab.objects.vulnerability_case import (
-        VulnerabilityCase,
-    )
-    from vultron.wire.as2.vocab.objects.vulnerability_report import (
-        VulnerabilityReport,
-    )
-
-    report = VulnerabilityReport(
-        name="CVE-2025-006",
-        content="Linked through ReportCaseLink",
-        attributed_to="https://example.org/finder",
-    )
-    case = VulnerabilityCase()
-
-    dl.create(report)
-    dl.save(case)
-    dl.save(VultronReportCaseLink(report_id=report.id_, case_id=case.id_))
-
-    result = dl.find_case_by_report_id(report.id_)
-
-    assert result is not None
-    assert isinstance(result, VulnerabilityCase)
-    assert result.id_ == case.id_
-
-
-# ---------------------------------------------------------------------------
-# File-backed integration test
-# ---------------------------------------------------------------------------
-
-
-@pytest.mark.integration
-def test_file_backed_store_and_retrieve(file_dl):
-    """Data written to a file-backed DataLayer is readable from the same URL."""
-    from vultron.wire.as2.vocab.base.objects.object_types import as_Note
-
-    note = as_Note(content="Integration test note")
-    file_dl.save(note)
-    result = file_dl.read(note.id_)
-    assert result is not None
-    assert result.id_ == note.id_  # type: ignore[union-attr]
-
-
-# ---------------------------------------------------------------------------
-# DL-REHYDRATE: semantic type recovery tests
-# ---------------------------------------------------------------------------
-
-
-class TestRehydrateFields:
-    """_rehydrate_fields expands dehydrated string IDs back to typed objects."""
-
-    def test_offer_object_field_expanded_to_vulnerability_report(self, dl):
-        """_RmSubmitReportActivity.object_ is a VulnerabilityReport after read."""
-        from vultron.wire.as2.vocab.objects.vulnerability_report import (
-            VulnerabilityReport,
-        )
-
-        report = VulnerabilityReport(name="CVE-TEST-001", content="Test body")
-        offer = rm_submit_report_activity(
-            report,
-            "https://alice.example.org",
-            actor="https://alice.example.org",
-        )
-        dl.save(report)
-        dl.save(offer)
-
-        result = dl.read(offer.id_)
-
-        assert isinstance(result, as_Offer)
-        assert isinstance(result.object_, VulnerabilityReport)  # type: ignore[union-attr]
-        assert result.object_.name == "CVE-TEST-001"  # type: ignore[union-attr]
-
-    def test_missing_nested_object_keeps_string(self, dl):
-        """When a referenced object is not in the DB, the string ID is kept."""
-        from vultron.wire.as2.vocab.objects.vulnerability_report import (
-            VulnerabilityReport,
-        )
-
-        report = VulnerabilityReport(name="CVE-MISSING", content="Body")
-        offer = rm_submit_report_activity(
-            report,
-            "https://alice.example.org",
-            actor="https://alice.example.org",
-        )
-        # Save offer but NOT the report — reference is dangling
-        dl.save(offer)
-
-        result = dl.read(offer.id_)
-
-        # Object field keeps the string ID since the report cannot be resolved
-        assert result is not None
-        assert isinstance(result.object_, str)  # type: ignore[union-attr]
-
-
-class TestCoerceToSemanticClass:
-    """_coerce_to_semantic_class promotes base-vocab activities to subtypes."""
-
-    def test_rm_submit_report_round_trip_returns_specific_class(self, dl):
-        """dl.read returns _RmSubmitReportActivity, not generic as_Offer."""
-        from vultron.wire.as2.vocab.objects.vulnerability_report import (
-            VulnerabilityReport,
-        )
-
-        report = VulnerabilityReport(name="CVE-ROUND-TRIP", content="Body")
-        offer = rm_submit_report_activity(
-            report,
-            "https://alice.example.org",
-            actor="https://alice.example.org",
-        )
-        dl.save(report)
-        dl.save(offer)
-
-        result = dl.read(offer.id_)
-
-        assert type(result).__name__ == "_RmSubmitReportActivity"
-
-    def test_em_propose_embargo_round_trip_returns_specific_class(self, dl):
-        """dl.read returns _EmProposeEmbargoActivity with EmbargoEvent object_."""
-        from vultron.wire.as2.vocab.objects.embargo_event import EmbargoEvent
-        from vultron.wire.as2.vocab.objects.vulnerability_case import (
-            VulnerabilityCase,
-        )
-
-        case = VulnerabilityCase()
-        embargo = EmbargoEvent(context=case.id_)
-        proposal = em_propose_embargo_activity(
-            embargo,
-            context=case.id_,
-            actor="https://alice.example.org",
-        )
-        dl.save(case)
-        dl.save(embargo)
-        dl.save(proposal)
-
-        result = dl.read(proposal.id_)
-
-        assert type(result).__name__ == "_EmProposeEmbargoActivity"
-        assert isinstance(result.object_, EmbargoEvent)  # type: ignore[union-attr]
-
-    def test_accept_invite_round_trip_returns_specific_class_from_generic_parse(
-        self, dl
-    ):
-        """Generic inbound Accept(Invite(...)) reads back as RmAcceptInviteToCaseActivity."""
-        from typing import cast
-
-        from vultron.wire.as2.parser import parse_activity
-        from vultron.wire.as2.vocab.base.objects.actors import as_Organization
-
-        parsed = cast(
-            as_Accept,
-            parse_activity(
-                {
-                    "type": "Accept",
-                    "id": "urn:uuid:accept-invite-roundtrip-1",
-                    "actor": "https://example.org/actors/coordinator",
-                    "inReplyTo": "urn:uuid:invite-roundtrip-1",
-                    "object": {
-                        "type": "Invite",
-                        "id": "urn:uuid:invite-roundtrip-1",
-                        "actor": "https://example.org/actors/vendor",
-                        "object": {
-                            "type": "Organization",
-                            "id": "https://example.org/actors/coordinator",
-                            "name": "Coordinator",
-                        },
-                        "target": {
-                            "type": "VulnerabilityCase",
-                            "id": "https://example.org/cases/case-roundtrip-1",
-                        },
-                        "to": ["https://example.org/actors/coordinator"],
-                    },
-                },
-            ),
-        )
-
-        nested_invite = parsed.object_
-        assert nested_invite is not None
-        dl.save(
-            as_Organization(
-                id_="https://example.org/actors/coordinator",
-                name="Coordinator",
-            )
-        )
-        dl.save(nested_invite)
-        dl.save(parsed)
-
-        result = dl.read(parsed.id_)
-
-        assert isinstance(result, as_Accept)
-        assert isinstance(result.object_, as_Invite)
-        assert result.object_.id_ == "urn:uuid:invite-roundtrip-1"
-        assert result.in_reply_to == "urn:uuid:invite-roundtrip-1"
-
-    def test_announce_log_entry_round_trip_returns_specific_class(self, dl):
-        """dl.read returns AnnounceLogEntryActivity with CaseLedgerEntry object_."""
-        from vultron.core.models.case_ledger import (
-            HashChainLedgerRecord,
-        )
-        from vultron.core.behaviors.sync.nodes.chain import (
-            _to_persistable_entry,
-        )
-        from vultron.wire.as2.vocab.objects.case_ledger_entry import (
-            CaseLedgerEntry as WireCaseLedgerEntry,
-        )
-
-        chain_entry = HashChainLedgerRecord(
-            case_id="https://example.org/cases/case-sync-1",
-            log_index=0,
-            object_id="https://example.org/activities/logged-1",
-            event_type="log_entry_committed",
-            payload_snapshot={"status": "ok"},
-            prev_log_hash=_ZERO_HASH,
-        )
-        entry = _to_persistable_entry(chain_entry)
-        announce = announce_log_entry_activity(
-            WireCaseLedgerEntry.from_core(entry),
-            actor="https://example.org/actors/case-actor",
-        )
-        dl.save(entry)
-        dl.save(announce)
-
-        result = dl.read(announce.id_)
-
-        assert isinstance(result, as_Announce)
-        assert isinstance(result.object_, WireCaseLedgerEntry)  # type: ignore[union-attr]
-        assert result.object_.case_id == entry.case_id  # type: ignore[union-attr]
-        assert result.object_.log_object_id == entry.log_object_id  # type: ignore[union-attr]
-
-    def test_non_activity_object_not_coerced(self, dl):
-        """Non-activity objects (e.g. VulnerabilityReport) are returned as-is."""
-        from vultron.wire.as2.vocab.objects.vulnerability_report import (
-            VulnerabilityReport,
-        )
-
-        report = VulnerabilityReport(name="CVE-PLAIN", content="Body")
-        dl.save(report)
-
-        result = dl.read(report.id_)
-
-        assert isinstance(result, VulnerabilityReport)
-
-
-class TestListMethod:
-    """dl.list_objects(type_key) returns fully rehydrated typed domain objects."""
-
-    def test_list_empty_for_unknown_type(self, dl):
-        """list() returns empty for a type with no stored records."""
-        result = dl.list_objects("NoSuchType")
-        assert result == []
-
-    def test_list_returns_saved_objects(self, dl):
-        """list() returns all objects of the requested type."""
-        from vultron.wire.as2.vocab.objects.vulnerability_report import (
-            VulnerabilityReport,
-        )
-
-        r1 = VulnerabilityReport(name="CVE-1", content="Body 1")
-        r2 = VulnerabilityReport(name="CVE-2", content="Body 2")
-        dl.save(r1)
-        dl.save(r2)
-
-        results = dl.list_objects("VulnerabilityReport")
-
-        assert len(results) == 2
-        ids = {obj.id_ for obj in results}
-        assert r1.id_ in ids
-        assert r2.id_ in ids
-
-    def test_list_filters_by_type(self, dl):
-        """list() only returns objects of the requested type, not others."""
-        from vultron.wire.as2.vocab.base.objects.object_types import as_Note
-        from vultron.wire.as2.vocab.objects.vulnerability_report import (
-            VulnerabilityReport,
-        )
-
-        report = VulnerabilityReport(name="CVE-FILTER", content="Body")
-        note = as_Note(content="Unrelated note")
-        dl.save(report)
-        dl.save(note)
-
-        reports = dl.list_objects("VulnerabilityReport")
-        notes = dl.list_objects("Note")
-
-        assert len(reports) == 1
-        assert reports[0].id_ == report.id_
-        assert len(notes) == 1
-        assert notes[0].id_ == note.id_
-
-    def test_list_returns_typed_objects_not_dicts(self, dl):
-        """list() returns domain model instances, not raw dict records."""
-        from vultron.wire.as2.vocab.objects.vulnerability_report import (
-            VulnerabilityReport,
-        )
-
-        report = VulnerabilityReport(name="CVE-TYPED", content="Body")
-        dl.save(report)
-
-        results = dl.list_objects("VulnerabilityReport")
-
-        assert len(results) == 1
-        assert isinstance(results[0], VulnerabilityReport)
-
-    def test_list_returns_semantic_class_for_activities(self, dl):
-        """list() applies semantic coercion so activities get their subtype."""
-        from vultron.wire.as2.vocab.base.objects.actors import as_Service
-        from vultron.wire.as2.vocab.objects.vulnerability_case import (
-            VulnerabilityCaseStub,
-        )
-
-        actor = as_Service(name="Coordinator")
-        invitee = as_Service(name="Vendor")
-        stub = VulnerabilityCaseStub(id_="https://example.org/cases/list-test")
-        invite = rm_invite_to_case_activity(
-            invitee,
-            target=stub,
-            actor=actor.id_,
-        )
-        dl.save(actor)
-        dl.save(invitee)
-        dl.save(invite)
-
-        results = dl.list_objects("Invite")
-
-        assert len(results) == 1
-        assert isinstance(results[0], as_Invite)
-
-
-# ---------------------------------------------------------------------------
-# hydrate() tests (CBT-05-005)
-# ---------------------------------------------------------------------------
-
-
-def test_hydrate_expands_list_ref_field(dl):
-    """hydrate() resolves case_participants string IDs to stored objects."""
-    from vultron.wire.as2.vocab.objects.vulnerability_case import (
-        VulnerabilityCase,
-    )
-    from vultron.wire.as2.vocab.objects.case_participant import CaseParticipant
-    from vultron.core.states.roles import CVDRole
-
-    case_actor_id = "https://example.org/actors/case-actor-hydrate"
-    case = VulnerabilityCase()
-
-    participant = CaseParticipant(
-        case_roles=[CVDRole.CASE_MANAGER],
-        attributed_to=case_actor_id,
-        context=case.id_,
-        name="test-participant",
-    )
-    dl.save(participant)
-
-    # Store case with participant as string ID (as it is stored in practice)
-    case.case_participants = [participant.id_]
-    dl.save(case)
-
-    # Read case back — still has string ID
-    stored_case = dl.read(case.id_)
-    assert stored_case is not None
-    assert isinstance(stored_case.case_participants[0], str)
-
-    # hydrate() should resolve the string to the full participant object
-    hydrated = dl.hydrate(stored_case)
-    assert hydrated is not stored_case  # new object returned
-    assert len(hydrated.case_participants) == 1
-    # dl.read reconstructs from storage as the base CaseParticipant type
-    from vultron.wire.as2.vocab.objects.case_participant import CaseParticipant
-
-    assert isinstance(hydrated.case_participants[0], CaseParticipant)
-    assert hydrated.case_participants[0].id_ == participant.id_
-
-
-def test_hydrate_leaves_already_expanded_participants_unchanged(dl):
-    """hydrate() leaves non-string participants unchanged."""
-    from vultron.wire.as2.vocab.objects.vulnerability_case import (
-        VulnerabilityCase,
-    )
-    from vultron.wire.as2.vocab.objects.case_participant import CaseParticipant
-    from vultron.core.states.roles import CVDRole
-
-    case_actor_id = "https://example.org/actors/case-actor-noop"
-    case = VulnerabilityCase()
-    participant = CaseParticipant(
-        case_roles=[CVDRole.CASE_MANAGER],
-        attributed_to=case_actor_id,
-        context=case.id_,
-        name="already-expanded",
-    )
-    dl.save(participant)
-    case.case_participants = [participant]
-    dl.save(case)
-
-    stored_case = dl.read(case.id_)
-    # Even if stored_case has string IDs after round-trip, hydrate must not fail
-    hydrated = dl.hydrate(stored_case)
-    assert len(hydrated.case_participants) == 1
-
-
-def test_hydrate_keeps_unresolvable_string_ids(dl):
-    """hydrate() keeps participant IDs that don't exist in the DataLayer."""
-    from vultron.wire.as2.vocab.objects.vulnerability_case import (
-        VulnerabilityCase,
-    )
-
-    missing_id = "urn:uuid:00000000-0000-0000-0000-000000000000"
-    case = VulnerabilityCase()
-    case.case_participants = [missing_id]
-    dl.save(case)
-
-    stored_case = dl.read(case.id_)
-    hydrated = dl.hydrate(stored_case)
-    assert hydrated.case_participants[0] == missing_id
-
-
-def test_hydrate_warns_for_unresolvable_string_ids(dl, caplog):
-    """hydrate() logs a WARNING when a list item ID cannot be resolved."""
-    import logging
-    from vultron.wire.as2.vocab.objects.vulnerability_case import (
-        VulnerabilityCase,
-    )
-
-    missing_id = "urn:uuid:00000000-0000-0000-0000-000000000001"
-    case = VulnerabilityCase()
-    case.case_participants = [missing_id]
-    dl.save(case)
-
-    stored_case = dl.read(case.id_)
-    with caplog.at_level(logging.WARNING):
-        dl.hydrate(stored_case)
-
-    assert any(
-        missing_id in record.message
-        for record in caplog.records
-        if record.levelname == "WARNING"
-    ), "Expected a WARNING log mentioning the unresolvable participant ID"
-
-
-# ---------------------------------------------------------------------------
-# enqueue_callback — outbox_append and record_outbox_item (AC-1)
-# ---------------------------------------------------------------------------
-
-
-@pytest.fixture
-def scoped_dl():
-    """Actor-scoped in-memory SqliteDataLayer for callback tests."""
-    instance = SqliteDataLayer(
-        "sqlite:///:memory:",
-        actor_id="https://example.org/alice",
-    )
-    yield instance
-    instance.clear_all()
-    instance.close()
-
-
-def test_outbox_append_calls_callback_with_actor_id(scoped_dl):
-    """outbox_append fires the enqueue_callback with the actor_id."""
-    calls: list[str] = []
-    scoped_dl.set_enqueue_callback(calls.append)
-    scoped_dl.outbox_append("urn:test:act-001")
-    assert calls == ["https://example.org/alice"]
-
-
-def test_outbox_append_no_callback_by_default():
-    """outbox_append does not raise when no callback is set (default)."""
-    dl = SqliteDataLayer(
-        "sqlite:///:memory:", actor_id="https://example.org/bob"
-    )
-    dl.outbox_append("urn:test:act-001")  # should not raise
-    dl.clear_all()
-    dl.close()
-
-
-def test_outbox_append_callback_cleared_by_set_none(scoped_dl):
-    """set_enqueue_callback(None) disables future notifications."""
-    calls: list[str] = []
-    scoped_dl.set_enqueue_callback(calls.append)
-    scoped_dl.set_enqueue_callback(None)
-    scoped_dl.outbox_append("urn:test:act-001")
-    assert calls == []
-
-
-def test_record_outbox_item_calls_callback_with_actor_id():
-    """record_outbox_item fires the enqueue_callback with the explicit actor_id."""
-    dl = SqliteDataLayer("sqlite:///:memory:")
-    calls: list[str] = []
-    dl.set_enqueue_callback(calls.append)
-    dl.record_outbox_item("https://example.org/carol", "urn:test:act-002")
-    assert calls == ["https://example.org/carol"]
-    dl.clear_all()
-    dl.close()
-
-
-def test_record_outbox_item_no_callback_by_default():
-    """record_outbox_item does not raise when no callback is set."""
-    dl = SqliteDataLayer("sqlite:///:memory:")
-    dl.record_outbox_item("https://example.org/carol", "urn:test:act-002")
-    dl.clear_all()
-    dl.close()
-
-
-def test_clone_for_actor_inherits_callback():
-    """clone_for_actor() inherits the parent's enqueue_callback."""
-    calls: list[str] = []
-    parent = SqliteDataLayer("sqlite:///:memory:")
-    parent.set_enqueue_callback(calls.append)
-    clone = parent.clone_for_actor("https://example.org/alice")
-    clone.outbox_append("urn:test:act-003")
-    assert calls == ["https://example.org/alice"]
-    parent.clear_all()
-    parent.close()
-
-
-def test_enqueue_callback_exception_does_not_propagate(scoped_dl):
-    """A callback that raises must not abort the outbox_append operation."""
-
-    def bad_callback(actor_id: str) -> None:
-        raise RuntimeError("callback failure")
-
-    scoped_dl.set_enqueue_callback(bad_callback)
-    # Should not raise; the item must still be enqueued
-    scoped_dl.outbox_append("urn:test:act-001")
-    assert "urn:test:act-001" in scoped_dl.outbox_list()
-
-
-def test_set_enqueue_callback_replaces_previous(scoped_dl):
-    """set_enqueue_callback replaces any previously registered callback."""
-    calls_a: list[str] = []
-    calls_b: list[str] = []
-    scoped_dl.set_enqueue_callback(calls_a.append)
-    scoped_dl.set_enqueue_callback(calls_b.append)
-    scoped_dl.outbox_append("urn:test:act-001")
-    assert calls_a == []
-    assert calls_b == ["https://example.org/alice"]
